Adiciona scrollbar

+++ window.py
@@ -7,38 +7,38 @@ class Window:
         self.y_max = y_max
 
     def move_up(self) -> None:
-        shift = int(0.2*abs(self.y_max - self.y_min))
+        shift = 0.2*abs(self.y_max - self.y_min)
         self.y_min += shift
         self.y_max += shift
 
     def move_down(self) -> None:
-        shift = int(0.2 * abs(self.y_max - self.y_min))
+        shift = 0.2 * abs(self.y_max - self.y_min)
         self.y_min -= shift
         self.y_max -= shift
 
     def move_left(self) -> None:
-        shift = int(0.2 * abs(self.x_max - self.x_min))
+        shift = 0.2 * abs(self.x_max - self.x_min)
         self.x_min -= shift
         self.x_max -= shift
 
     def move_right(self) -> None:
-        shift = int(0.2 * abs(self.x_max - self.x_min))
+        shift = 0.2 * abs(self.x_max - self.x_min)
         self.x_min += shift
         self.x_max += shift
 
     def zoom_in(self) -> None:
-        x_dif = int(0.1 * abs(self.x_max - self.x_min))
+        x_dif = 0.1 * abs(self.x_max - self.x_min)
         self.x_min += x_dif
         self.x_max -= x_dif
-        y_dif = int(0.1 * abs(self.y_max - self.y_min))
+        y_dif = 0.1 * abs(self.y_max - self.y_min)
         self.y_min += y_dif
         self.y_max -= y_dif
 
     def zoom_out(self) -> None:
-        x_dif = int(0.1 * abs(self.x_max - self.x_min))
+        x_dif = 0.1 * abs(self.x_max - self.x_min)
         self.x_min -= x_dif
         self.x_max += x_dif
-        y_dif = int(0.1 * abs(self.y_max - self.y_min))
+        y_dif = 0.1 * abs(self.y_max - self.y_min)
         self.y_min -= y_dif
         self.y_max += y_dif
